Implementing new state switching architecture.
git-svn-id: file:///home/chad/tmp/glscry/trunk/glscry@122 a68c7e1a-86db-0310-a1c4-e68a6bf46cff

+++ src/GeometryTest.h
@@ -81,7 +81,7 @@ namespace scry {
             }
         }
 
-        size_t getScreenCoverage() const {
+        Uint64 getScreenCoverage() const {
             return _screenCoverage;
         }
 
@@ -97,7 +97,7 @@ namespace scry {
             const char* name, PumpGetter getter);
 
         Inited<size_t, 4096> _batchSize;
-        Zeroed<size_t> _screenCoverage;
+        Zeroed<Uint64> _screenCoverage;
 
         /// Size of vertex arrays when using indexed geometry.
         Zeroed<size_t> _vertexArraySize;  // In 'datatype's
